moved py into check
