+++ build.sh
@@ -8,4 +8,4 @@ eval $(go env)
 export GO15VENDOREXPERIMENT="1"
 
 go build -o ${BIN_PATH}/agent ${REPO_PATH}/cmd/agent
-cp -r ./pkg/py/dist/ ${BIN_PATH}/dist/
+cp -r ./pkg/collector/check/py/dist/ ${BIN_PATH}/dist/
