+++ pkg/collector/check/py/config.go
