+++ pkg/collector/check/py/config_test.go
