+++ pkg/collector/check/py/loader.go
