+++ pkg/collector/check/py/loader_test.go
