+++ pkg/collector/check/py/tests/bad.yaml
