+++ pkg/collector/check/py/tests/complex.yaml
