+++ pkg/collector/check/py/tests/testcheck.py
