+++ pkg/collector/check/py/tests/testcheck.yaml
