+++ pkg/collector/check/py/tests/testcheck2.py
