+++ pkg/collector/check/py/utils.go
