+++ pkg/collector/check/py/utils_test.go
